fix(island jobs): fixed bug where completing site jobs resulted in an NullPointerException

+++ src/main/java/de/uniks/stp24/component/game/jobs/IslandOverviewJobProgressComponent.java
@@ -90,7 +90,7 @@ public class IslandOverviewJobProgressComponent extends Pane implements Reusable
         }
 
         this.jobPositionText.setText(systemJobs.indexOf(job)+1 + ".");
-        this.jobTimeRemaining.setText(String.format("%s/%s", job.progress(), job.total()));
+        this.jobTimeRemaining.setText(String.format("%s/%s", job.progress(), (int) job.total()));
 
         this.resourceObservableList.clear();
         job.cost().forEach((name, count) -> this.resourceObservableList
